[fix] layout api

+++ src/app/modules/layout/layout.service.ts
@@ -1,15 +1,19 @@
 import ApiError from "../../errorHandlers/ApiError";
+import { deleteFile } from "../../helper/deleteFile";
 import { ILayout } from "./layout.inteface";
 import LayoutModel from "./layout.model";
 
 const createLayoutIntodb = async (payload: ILayout): Promise<ILayout> => {
+  const layouts = await LayoutModel.countDocuments();
+  if (layouts < 0) {
+    payload.selected = true;
+  }
+
   return await LayoutModel.create(payload);
 };
 
-const getLayoutFromdb = async (selected?: string): Promise<ILayout[]> => {
-  const isSelected = Boolean(selected);
-
-  const layouts = await LayoutModel.find({ selected: isSelected });
+const getLayoutFromdb = async (): Promise<ILayout[]> => {
+  const layouts = await LayoutModel.find().sort({ selected: -1 });
   if (layouts.length < 0) {
     throw new ApiError(404, "Layout Not Found");
   }
@@ -27,10 +31,17 @@ const getLayoutById = async (id: string): Promise<ILayout> => {
 };
 
 const updateLayoutIntodb = async (
-  payload: Record<string, unknown>,
+  payload: Partial<ILayout>,
   id: string
 ): Promise<ILayout> => {
-  const layout = await LayoutModel.findByIdAndUpdate(id, payload);
+  if (payload.selected) {
+    await unselectCurrentLayout();
+  }
+
+  const layout = await LayoutModel.findByIdAndUpdate(id, payload, {
+    new: true,
+    runValidators: true,
+  });
   if (!layout) {
     throw new ApiError(404, "Layout not found");
   }
@@ -39,14 +50,25 @@ const updateLayoutIntodb = async (
 };
 
 const deleteLayoutFromdb = async (id: string) => {
-  const layout = await LayoutModel.findByIdAndDelete(id, {
-    runValidators: true,
-    new: true,
-  });
+  const layout = await LayoutModel.findById(id);
 
   if (!layout) {
     throw new ApiError(404, "Layout not found");
   }
+  if (layout.selected) {
+    throw new ApiError(401, "Not Able to Delete Selected Layout");
+  }
+
+  await LayoutModel.findByIdAndDelete(id, {
+    new: true,
+    runValidators: true,
+  });
+
+  await deleteFile(layout.image);
+};
+
+const unselectCurrentLayout = async () => {
+  await LayoutModel.findOneAndUpdate({ selected: true }, { selected: false });
 };
 
 export const layoutServices = {
